AGV generator updated with correct wheelTemperature

+++ Producer/src/main/java/pl/bartryb/producer/models/AutonomousGuidedVehicle.java
@@ -31,5 +31,5 @@ public class AutonomousGuidedVehicle {
     public Integer machineId;
 
     @JsonProperty
-    public Integer wheelsTemperature;
+    public Double wheelsTemperature;
 }
